refactor: Inline editor chooser modal content
Remove the separate testB component now that the A/B test is over.
